refactor: extract billing accordion
